Fixed problems related to precision being int, short, and unsigned short in various places while PREC field is short.  Changed nullRecord to set precision to 0 instead of -1.  Was clobbering stack when localCvtDoubleToString got called when precision was -1, which was converted to 65535U.  Also fixed menuDraw and buttonDraw to return if widget is null.  These files may also have changes related to WIN32. WIN32 is still being debugged.  These changes fix part of the WIN32 problems.

+++ medm/medmBar.c
@@ -267,7 +267,7 @@ static void barUpdateGraphicalInfoCb(XtPointer cd) {
     Bar *pb = (Bar *) pd->clientData;
     XcVType hopr, lopr, val;
     Pixel pixel;
-    int precision;
+    short precision;
     Widget widget = pb->dlElement->widget;
     DlBar *dlBar = pb->dlElement->structure.bar;
 
@@ -308,7 +308,7 @@ static void barUpdateGraphicalInfoCb(XtPointer cd) {
       XcNlowerBound,lopr.lval,
       XcNupperBound,hopr.lval,
       XcNbarForeground,pixel,
-      XcNdecimals, precision,
+      XcNdecimals, (int)precision,
       NULL);
     XcBGUpdateValue(widget,&val);
 }
